code/src/main/java/ru/merkurev/leetcode/problems/p0021/Solution.java
package ru.merkurev.leetcode.problems.p0021;

/**
 * You are given the heads of two sorted linked lists list1 and list2.
 * <p>
 * Merge the two lists into one sorted list. The list should be made by splicing together the nodes of the first two lists.
 * <p>
 * Return the head of the merged linked list.
 * <p>
 * Constraints:
 * <p>
 * • The number of nodes in both lists is in the range [0, 50].<br>
 * • -100 <= Node.val <= 100<br>
 * • Both list1 and list2 are sorted in non-decreasing order.<br>
 */
class Solution {
    public ListNode mergeTwoLists(ListNode list1, ListNode list2) {

        if (list1 == null) {
            return list2;
        }
        if (list2 == null) {
            return list1;
        }

        ListNode head;
        if (list1.val <= list2.val) {
            head = list1;
            list1 = list1.next;
        } else {
            head = list2;
            list2 = list2.next;
        }

        ListNode current = head;
        while (list1 != null && list2 != null) {
            if (list1.val <= list2.val) {
                current.next = list1;
                current = list1;
                list1 = list1.next;
            } else {
                current.next = list2;
                current = list2;
                list2 = list2.next;
            }
        }
        current.next = list1 == null ? list2 : list1;
        return head;
    }
}
